Different Sectors
want to see improvement in the terrain of the sectors in the game. Add mountains like level, desert and
forest sectors. Also, I want water flow and direction. It is like a conveyor belt with a certain speed and direction on a block of water. it would change the mechanics of the game. Serpulo has a flaw: even though the planet is changing day and night, the planet itself, or the star, does not move. I speak Russian, so i used Google translate. If you don't understand me, send me an email. I know, that you (maybe) understand Russian. Good luck!